Js reconfigured
