+++ package.json
@@ -21,7 +21,10 @@
         "vue-template-compiler": "^2.6.10"
     },
     "dependencies": {
+        "gsap": "^3.5.1",
+        "jquery": "^3.5.1",
         "numeral": "^2.0.6",
+        "sticky-kit": "^1.1.3",
         "vue-avatar": "^2.1.8"
     }
 }
