+++ public/mix-manifest.json
@@ -1,4 +1,8 @@
 {
     "/js/app.js": "/js/app.js",
-    "/css/app.css": "/css/app.css"
+    "/css/app.css": "/css/app.css",
+    "/js/lazyload_users.js": "/js/lazyload_users.js",
+    "/js/sticky-kit_users.js": "/js/sticky-kit_users.js",
+    "/js/jquery.min.js": "/js/jquery.min.js",
+    "/js/sticky-kit.min.js": "/js/sticky-kit.min.js"
 }
