+++ resources/js/lazyload_users.js
@@ -1,0 +1,3 @@
+$(document).ready(function() {
+lazyload();
+});
