+++ resources/js/sticky-kit_users.js
@@ -1,0 +1,3 @@
+$(document).ready(function() {
+$(".sidebar").stick_in_parent({ offset_top: 120 });
+});
