+++ resources/sass/_opensans.scss
@@ -1,0 +1,4 @@
+@font-face {
+    font-family: "Open Sans";
+    src: url("../fonts/OpenSans-Regular.ttf") format("truetype");
+}
