+++ resources/sass/app.scss
@@ -1,6 +1,6 @@
 /* Fonts */
 
-@import url("https://fonts.googleapis.com/css?family=Open+Sans&display=swap");
+@import "_opensans.scss";
 
 /* Files */
 @import "general";
