+++ resources/views/layouts/dashboard.blade.php
@@ -179,7 +179,7 @@
          <!-- /.Footer-->
          
          <!--Scripts -->
-        <script src="{{ asset('js/jquery3.4.1.min.js') }}"></script>
+        <script src="{{ asset('js/jquery.min.js') }}"></script>
         @stack('scripts')
         <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.1.1/gsap.min.js"></script>
         <script src="{{ asset('js/app.js') }}" defer></script>
